+++ internal/command/select_test.go
@@ -441,6 +441,15 @@ d.e.f`)),
 		}
 	})
 
+	t.Run("Issue 392 panic", runTest(
+		[]string{"-r", "csv", "--csv-comma", ";", "-w", "json", "equal([], )"},
+		[]byte(`Hello;There;
+1;2;`),
+		[]byte("false\n"),
+		nil,
+		nil,
+	))
+  
 	t.Run("Issue346", func(t *testing.T) {
 		t.Run("Select null or default string", runTest(
 			[]string{"-r", "json", "orDefault(foo,string(nope))"},