+++ value.go
@@ -302,7 +302,8 @@ func (v Value) Index(i int) Value {
 func (v Value) Append() Value {
 	currentLen := v.Len()
 	newLen := currentLen + 1
-	updatedSlice := reflect.MakeSlice(sliceInterfaceType, newLen, newLen)
+
+	updatedSlice := reflect.MakeSlice(reflect.TypeOf(v.Interface()), newLen, newLen)
 	// copy all existing elements into updatedSlice.
 	// this leaves the last element empty.
 	for i := 0; i < currentLen; i++ {
@@ -311,18 +312,22 @@ func (v Value) Append() Value {
 		)
 	}
 
-	v.FirstAddressable().Set(updatedSlice)
+	firstAddressable := v.FirstAddressable()
+	firstAddressable.Set(updatedSlice)
 
+	// This code was causing a panic...
+	// It doesn't seem necessary. Leaving here for reference in-case it was needed.
+	// See https://github.com/TomWright/dasel/issues/392
 	// Set the last element to uninitialised.
-	updatedSlice.Index(currentLen).Set(
-		v.Index(currentLen).asUninitialised().Value,
-	)
+	//updatedSlice.Index(currentLen).Set(
+	//	v.Index(currentLen).asUninitialised().Value,
+	//)
 
 	return v
 }
 
-var sliceInterfaceType = reflect.TypeOf([]interface{}{})
-var mapStringInterfaceType = reflect.TypeOf(map[string]interface{}{})
+var sliceInterfaceType = reflect.TypeFor[[]any]()
+var mapStringInterfaceType = reflect.TypeFor[map[string]interface{}]()
 
 var UninitialisedPlaceholder interface{} = "__dasel_not_found__"